Refactoring client/navigation to use new conventions
