+++ client/navigation/components/category-title/category-title.scss
