+++ client/navigation/components/favorite-button/favorite-button.scss
