+++ client/navigation/index.js
@@ -6,7 +6,7 @@ import { withNavigationHydration } from '@woocommerce/data';
 /**
  * Internal dependencies
  */
-import './style.scss';
+import './navigation.scss';
 import Container from './components/container';
 
 const HydratedNavigation = withNavigationHydration( window.wcNavigation )(
